make viewcoords() return ints by default

+++ doc/examples_src/somex_pynd/00003.py
@@ -1,15 +1,15 @@
 def testcolors():
     xmin, ymin, xmax, ymax = viewcoords()
 
-    for y in range(int(ymin), int(ymin/3)):
+    for y in range(ymin, ymin/3):
         color('rlight')
         goto(xmin, y)
         lineto(xmax, y)
-    for y in range(int(ymin/3), int(ymax/3)):
+    for y in range(ymin/3, ymax/3):
         color('rmedium')
         goto(xmin, y)
         lineto(xmax, y)
-    for y in range(int(ymax/3), int(ymax)):
+    for y in range(ymax/3, ymax):
         color('rdark')
         goto(xmin, y)
         lineto(xmax, y)
